// frontend/src/app/core/api/model/vendorRatingDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VendorRatingDto { 
    communicationScore?: number;
    deliveryScore?: number;
    qualityScore?: number;
}